tt_beacon: do not terminate on send failure after previous success
Failure to send a beacon for any reason causes tt_beacon to exit.
This behavior forces a restart of the program when anything that
can trigger send errors occurs, even if it is as simple as a wire
being unplugged for a few seconds or an interface being restarted
in a network configuration script.

Instead of terminating, this change makes tt_beacon warn the user
one time via stderr that send failure occurred and the program
sleeps longer until sending succeeds again.

If the program fails to send a packet when started, it will still
terminate with an error message. Only one warning is emitted per
group of failures. When sending succeeds again, the warning switch
is reset.