tss_client: Auto-detect TSS server instead of hard-coding it
Code for autodetection of the TSS server by waiting for a beacon
was in tss_client but was never called from anywhere. A legacy IP
address was removed from a TSS server instance and turned out to
be required to use the server solely because of this missing link.
This change causes the TSS server auto-detect code to be called in
places that require knowledge of the server and eliminates most of
the hard-coded parts until all else fails.

The hard-coded server stuff needs to be pushed to a .conf file in
the future, but this will do for now.