tt_driverdrop: fix bug where USB PID was ignored for driver choice
A lot of weird drivers were being selected based on the USB VID
alone being a match. If the USB device doesn't match on class
parameters, matching instead must rely on the VID/PID both being
identical. This commit fixes this problem as simply as possible.

This also adds some ACPI devices to the blacklist and adds an
indicator for whether rsync or cp is being used to copy drivers.