Remove cat commands where unneeded, make backup read-only work
Backups were not mounting NTFS-3G read-only correctly. Since we
only tend to need it for NTFS right now, switch the mount command
to explicitly call mount.ntfs-3g -o ro.

Removed the use of 'cat' where it was not required, specifically
for input to 'tr' and 'grep' and such. Untested but probably not
going to cause any issues.