tss__common: stop SMART error/info showing when not appropriate
Some SSDs don't support certain SMART logging and this was
causing them to show SELE when there is actually no log to have
entries in in the first place. Also, the USB SAT string should
only be added if it's not blank; the unconditional adding caused
later empty string tests to see a "non-empty" string and show a
totally blank SMART line instead of no SMART line as expected.